replace int by long and excepties en kleine aanpassingen

+++ railway-app-station/src/main/java/com/railway/station_service/PlatformRepository.java
@@ -2,7 +2,7 @@ package com.railway.station_service;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlatformRepository  extends CrudRepository<Platform, Integer>  {
+public interface PlatformRepository  extends CrudRepository<Platform, Long>  {
 
 	
 }
